LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/dynamic-import/string-compilation-nonce-classic-expected.txt

CONSOLE MESSAGE: The source list for Content Security Policy directive 'script-src' contains an invalid source: ''unsafe-hashes''. It will be ignored.
CONSOLE MESSAGE: Refused to load http://localhost:8800/html/semantics/scripting-1/the-script-element/module/imports-a.js?label=setTimeout because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to execute a script for an inline event handler because 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to execute a script for an inline event handler because 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.

FAIL setTimeout must inherit the nonce from the triggering script, thus execute promise_test: Unhandled rejection with value: object "TypeError: Importing a module script failed."
PASS direct eval must inherit the nonce from the triggering script, thus execute
PASS indirect eval must inherit the nonce from the triggering script, thus execute
PASS the Function constructor must inherit the nonce from the triggering script, thus execute
FAIL reflected inline event handlers must not inherit the nonce from the triggering script, thus fail dummyDiv.onclick is not a function. (In 'dummyDiv.onclick()', 'dummyDiv.onclick' is null)
FAIL inline event handlers triggered via UA code must not inherit the nonce from the triggering script, thus fail assert_equals: the browser must be able to parse a string containing the import() syntax into a function expected "function" but got "object"
